extract funcs to file_validator.go
