+++ cmd/api/src/services/ingest/stream_decoder.go
